split post component
split post component into two separate sub-components
