+++ src/components/pagesComponents/posts/post/postContent/PostContet.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import classes from './postContent.module.scss';
+
+const PostContent = ({ post }) => {
+  return (
+    <div>
+      <p className={classes.text} data-test='text'>
+        {post.text}
+      </p>
+      {post.url && <img src={post.url} className={classes.postImage} data-test='post-image' alt='post picture' />}
+    </div>
+  );
+};
+
+export default PostContent;
