+++ src/components/pagesComponents/posts/post/postContent/postContent.module.scss
@@ -1,0 +1,31 @@
+.text {
+  font-size: 13px;
+  color: #333;
+}
+
+.postImage {
+  margin-top: 0.65em;
+  width: 100%;
+}
+
+@media (min-width: 768px) and (orientation: portrait) {
+  .text {
+    font-size: 20px;
+  }
+}
+
+@media (min-width: 700px) and(orientation:landscape) {
+  .text {
+    font-size: 14px;
+  }
+}
+
+@media (min-width: 1000px) and(orientation:landscape) {
+  .postImage {
+    margin-top: 0.7em;
+  }
+
+  .text {
+    font-size: 16px;
+  }
+}
